refactor(maps): extract MapDrawingLayer (drawing render + selection)
Move the drawing SVG overlay render (renderDrawing for pencil/rect/ellipse incl.
the pencil hit-area), plus the selected-drawing bounding box + corner resize
handle, into a presentational MapDrawingLayer component. Interaction handlers
(begin move/resize) and the drag/erase state stay in ActiveMapStage and are
passed as props. No behavior change; testids preserved and verified by the
drawing + drawing-permissions e2e specs (21 tests).
